refactor: ArticleValidator삭제 후 관련 클래스 수정
