+++ src/main/java/com/mobble/mobbleserver/application/comment/service/CommentQueryService.java
@@ -1,12 +1,14 @@
 package com.mobble.mobbleserver.application.comment.service;
 
+import com.mobble.mobbleserver.application.article.port.required.ArticleReadPort;
 import com.mobble.mobbleserver.application.comment.port.provided.CommentQueryPort;
 import com.mobble.mobbleserver.application.comment.port.required.CommentReadPort;
+import com.mobble.mobbleserver.domain.article.Article;
 import com.mobble.mobbleserver.domain.comment.Comment;
+import com.mobble.mobbleserver.global.exception.common.DomainException;
+import com.mobble.mobbleserver.global.exception.errorCode.article.ArticleErrorCode;
 import com.mobble.mobbleserver.infrastructure.persistence.comment.projection.CommentLikeInfoDto;
 import com.mobble.mobbleserver.infrastructure.web.comment.dto.response.RootCommentResponseDto;
-import com.mobble.mobbleserver.refactor.article.entity.Article;
-import com.mobble.mobbleserver.refactor.article.validator.ArticleValidator;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -20,12 +22,14 @@ import java.util.stream.Stream;
 @Transactional(readOnly = true)
 public class CommentQueryService implements CommentQueryPort {
 
-    private final ArticleValidator articleValidator;
     private final CommentReadPort commentReadPort;
+    private final ArticleReadPort articleReadPort;
 
     @Override
     public List<RootCommentResponseDto> getCommentListByArticle(Long articleId, Long memberId) {
-        Article article = articleValidator.findArticleByArticleIdOrThrow(articleId);
+        Article article = articleReadPort.findById(articleId)
+                .orElseThrow(() -> new DomainException(ArticleErrorCode.NOT_FOUND));
+
         List<Comment> comments = commentReadPort.findCommentsWithRepliesByArticleId(article.getId());
         Map<Long, CommentLikeInfoDto> likeInfoMap = getCommentLikeInfo(comments, memberId);
